+++ src/Catalog/Product.php
@@ -120,6 +120,12 @@ final class Product
     /** @var bool */
     protected $isMvp;
 
+    /** @var null|bool */
+    private $isSubscription;
+
+    /** @var null|bool */
+    private $isRenewable;
+
     /**
      * @internal
      *
@@ -156,12 +162,17 @@ final class Product
         $this->categoryPath = array_map(static function (array $category) : ProductCategory {
             return new ProductCategory($category);
         }, $data['categoryPath']);
+        $this->isSubscription = $data['isSubscription'] ?? null;
+        $this->isRenewable = $data['isRenewable'] ?? null;
         $this->declinations = array_map(function (array $declination) : Declination {
             // Only available for a product
             // For a MVP we're not able to know the shippings
             if (false === $this->isMvp()) {
                 $declination['shippings'] = $this->shippings;
             }
+
+            $declination['isSubscription'] = $this->isSubscription();
+            $declination['isRenewable'] = $this->isRenewable();
 
             return new Declination($declination);
         }, $data['declinations']);
@@ -531,5 +542,21 @@ final class Product
     public function isMvp(): bool
     {
         return $this->isMvp;
+    }
+
+    /**
+     * @return null|bool
+     */
+    public function isSubscription(): ?bool
+    {
+        return $this->isSubscription;
+    }
+
+    /**
+     * @return null|bool
+     */
+    public function isRenewable(): ?bool
+    {
+        return $this->isRenewable;
     }
 }